Move styled elements form OrderRow component to a separated file
