+++ src/cow-react/modules/limitOrders/pure/Orders/OrderRow/index.tsx
@@ -1,0 +1,190 @@
+import { useContext } from 'react'
+import { DefaultTheme, StyledComponent, ThemeContext } from 'styled-components/macro'
+import { OrderStatus } from 'state/orders/actions'
+import { Currency, CurrencyAmount } from '@uniswap/sdk-core'
+import { RateInfo } from '@cow/common/pure/RateInfo'
+import { MouseoverTooltipContent } from 'components/Tooltip'
+import { Trash2 } from 'react-feather'
+import { OrderParams } from '../utils/getOrderParams'
+import { getSellAmountWithFee } from '@cow/modules/limitOrders/utils/getSellAmountWithFee'
+import AlertTriangle from 'assets/cow-swap/alert.svg'
+import SVG from 'react-inlinesvg'
+import { TokenSymbol } from '@cow/common/pure/TokenSymbol'
+import { TokenAmount } from '@cow/common/pure/TokenAmount'
+import CurrencyLogo from 'components/CurrencyLogo'
+import useTimeAgo from 'hooks/useTimeAgo'
+import { ParsedOrder } from '@cow/modules/limitOrders/containers/OrdersWidget/hooks/useLimitOrdersList'
+import { OrderStatusBox } from '@cow/modules/limitOrders/pure/OrderStatusBox'
+import * as styledEl from './styled'
+
+export const orderStatusTitleMap: { [key in OrderStatus]: string } = {
+  [OrderStatus.PENDING]: 'Open',
+  [OrderStatus.PRESIGNATURE_PENDING]: 'Signing',
+  [OrderStatus.FULFILLED]: 'Filled',
+  [OrderStatus.EXPIRED]: 'Expired',
+  [OrderStatus.CANCELLED]: 'Cancelled',
+  [OrderStatus.CREATING]: 'Creating',
+  [OrderStatus.FAILED]: 'Failed',
+}
+
+const TIME_AGO_UPDATE_INTERVAL = 3000
+
+function CurrencyAmountItem({ amount }: { amount: CurrencyAmount<Currency> }) {
+  return (
+    <styledEl.AmountItem title={amount.toExact() + ' ' + amount.currency.symbol}>
+      <TokenAmount amount={amount} tokenSymbol={amount.currency} />
+    </styledEl.AmountItem>
+  )
+}
+
+function CurrencySymbolItem({ amount }: { amount: CurrencyAmount<Currency> }) {
+  return <CurrencyLogo currency={amount.currency} size="28px" />
+}
+
+const BalanceWarning = (symbol: string) => (
+  <styledEl.WarningParagraph>
+    <h3>Insufficient balance for this limit order</h3>
+    <p>
+      Your wallet currently has insufficient{' '}
+      <strong>
+        <TokenSymbol token={{ symbol }} />
+      </strong>{' '}
+      balance to execute this order.
+      <br />
+      The order is still open and will become executable when you top up your{' '}
+      <strong>
+        <TokenSymbol token={{ symbol }} />
+      </strong>{' '}
+      balance.
+    </p>
+  </styledEl.WarningParagraph>
+)
+
+const AllowanceWarning = (symbol: string) => (
+  <styledEl.WarningParagraph>
+    <h3>Insufficient approval for this limit order</h3>
+    <p>
+      This order is still open and valid, but you haven’t given CoW Swap sufficient allowance to spend{' '}
+      <strong>
+        <TokenSymbol token={{ symbol }} />
+      </strong>
+      .
+      <br />
+      The order will become executable when you approve{' '}
+      <strong>
+        <TokenSymbol token={{ symbol }} />
+      </strong>{' '}
+      in your account token page.
+    </p>
+  </styledEl.WarningParagraph>
+)
+
+export interface OrderRowProps {
+  order: ParsedOrder
+  RowElement: StyledComponent<'div', DefaultTheme>
+  isRateInversed: boolean
+  orderParams: OrderParams
+  onClick: () => void
+  getShowCancellationModal(order: ParsedOrder): (() => void) | null
+}
+
+export function OrderRow({
+  order,
+  RowElement,
+  isRateInversed,
+  getShowCancellationModal,
+  orderParams,
+  onClick,
+}: OrderRowProps) {
+  const { buyAmount, rateInfoParams, hasEnoughAllowance, hasEnoughBalance } = orderParams
+  const { parsedCreationTime, expirationTime } = order
+
+  const showCancellationModal = getShowCancellationModal(order)
+
+  const withWarning = !hasEnoughBalance || !hasEnoughAllowance
+  const theme = useContext(ThemeContext)
+
+  const expirationTimeAgo = useTimeAgo(expirationTime, TIME_AGO_UPDATE_INTERVAL)
+  const creationTimeAgo = useTimeAgo(parsedCreationTime, TIME_AGO_UPDATE_INTERVAL)
+
+  return (
+    <RowElement onClick={onClick}>
+      {/* Order sell/buy tokens */}
+      <styledEl.CurrencyCell>
+        <styledEl.CurrencyLogoPair>
+          <CurrencySymbolItem amount={getSellAmountWithFee(order)} />
+          <CurrencySymbolItem amount={buyAmount} />
+        </styledEl.CurrencyLogoPair>
+        <styledEl.CurrencyAmountWrapper>
+          <CurrencyAmountItem amount={getSellAmountWithFee(order)} />
+          <CurrencyAmountItem amount={buyAmount} />
+        </styledEl.CurrencyAmountWrapper>
+      </styledEl.CurrencyCell>
+
+      {/* Limit price */}
+      <styledEl.CellElement>
+        <styledEl.RateValue>
+          <RateInfo prependSymbol={false} noLabel={true} isInversed={isRateInversed} rateInfoParams={rateInfoParams} />
+        </styledEl.RateValue>
+      </styledEl.CellElement>
+
+      {/* Est. execution price */}
+      {/* Market price */}
+      <styledEl.CellElement doubleRow>
+        <b>1534.62 USDC</b>
+        <i>1531.33 USDC</i>
+      </styledEl.CellElement>
+
+      {/* Expires */}
+      {/* Created */}
+      <styledEl.CellElement doubleRow>
+        <b>{expirationTimeAgo}</b>
+        <i>{creationTimeAgo}</i>
+      </styledEl.CellElement>
+
+      {/* Filled % */}
+      <styledEl.CellElement doubleRow>
+        <b>13.12%</b>
+        <styledEl.ProgressBar value={13.12}></styledEl.ProgressBar>
+      </styledEl.CellElement>
+
+      <styledEl.CellElement>
+        <styledEl.StatusBox>
+          <OrderStatusBox cancelling={!!order.isCancelling} status={order.status} withWarning={withWarning}>
+            {order.isCancelling ? 'Cancelling...' : orderStatusTitleMap[order.status]}
+          </OrderStatusBox>
+          {withWarning && (
+            <styledEl.WarningIndicator>
+              <MouseoverTooltipContent
+                wrap={false}
+                bgColor={theme.alert}
+                content={
+                  <styledEl.WarningContent>
+                    {!hasEnoughBalance && BalanceWarning(order.inputToken.symbol || '')}
+                    {!hasEnoughAllowance && AllowanceWarning(order.inputToken.symbol || '')}
+                  </styledEl.WarningContent>
+                }
+                placement="bottom"
+              >
+                <SVG src={AlertTriangle} description="Alert" width="14" height="13" />
+              </MouseoverTooltipContent>
+            </styledEl.WarningIndicator>
+          )}
+        </styledEl.StatusBox>
+      </styledEl.CellElement>
+      <styledEl.CellElement>
+        {showCancellationModal && (
+          <styledEl.CancelOrderBtn
+            title="Cancel order"
+            onClick={(event) => {
+              event.stopPropagation()
+              showCancellationModal()
+            }}
+          >
+            <Trash2 size={16} />
+          </styledEl.CancelOrderBtn>
+        )}
+      </styledEl.CellElement>
+    </RowElement>
+  )
+}
